Combined Devoxx19ShowTopRatedToday/Week into Devoxx19ShowTopRated. Differences handled via mandatory configuration field topVotedType.
fixes TweetWallFX/TweetwallFX-Devoxx-2019-BE#12
